+++ TheGuideToTheNewEden/TheGuideToTheNewEden.Core/Models/PlanetColony/PlanetColonyDetail.cs
@@ -3,9 +3,9 @@ namespace TheGuideToTheNewEden.Core.Models.PlanetColony
 {
     public class PlanetColonyDetail
     {
-        public int PlanetId { get; set; }
-        public int OwnerId { get; set; }
-        public int UpgradeLevel { get; set; }
+        public long PlanetId { get; set; }
+        public long OwnerId { get; set; }
+        public long UpgradeLevel { get; set; }
         public List<PlanetPin> Pins { get; set; } = new List<PlanetPin>();
         public List<PlanetRoute> Routes { get; set; } = new List<PlanetRoute>();
         public List<PlanetLink> Links { get; set; } = new List<PlanetLink>();
